requirements.txt
Django==4.2
Pillow==11.2.1
django-phonenumber-field[phonenumbers]==8.0.0
